+++ History.md
@@ -1,3 +1,10 @@
+Unpublished
+===========
+
+  * update dependencies (nan v2.11.1)
+  * adapt to V8 7.0: replace v8Value->Uint32Value() with Nan::To<uint32_t>(v8Value).FromJust()
+  * adapt to V8 7.0: replace v8Value->BooleanValue() with Nan::To<bool>(v8Value).FromJust()
+
 2.0.3 / Jul 28 2018
 ===================